Invalidate mocks and affected files on change
